pyproject.toml
[project]
name = "aibuddy"
version = "0.1.0"
description = ""
authors = [
    {name = "Jace Trent",email = "trentjace58@gmail.com"}
]
requires-python = ">=3.12"
dependencies = [
    "django (>=6.0.2,<7.0.0)",
    "environs (>=14.6.0,<15.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "openai (>=2.24.0,<3.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode=false

